+++ src/main/java/com/b1n_ry/yigd/mixin/LivingEntityMixin.java
@@ -11,7 +11,7 @@ import org.spongepowered.asm.mixin.injection.At;
 import org.spongepowered.asm.mixin.injection.Inject;
 import org.spongepowered.asm.mixin.injection.callback.CallbackInfo;
 
-@Mixin(LivingEntity.class)
+@Mixin(value = LivingEntity.class, priority = 500)
 public class LivingEntityMixin {
     @Inject(method = "drop", at = @At("HEAD"))
     private void drop(DamageSource damageSource, CallbackInfo ci) {